Delta time

+++ Shoot The Komsai/komsai.cpp
@@ -1,15 +1,15 @@
 #include "komsai.h"
 
-Komsai::Komsai() : komsaiY(0), komsaiX(0), komsaiSpeed(5), komsaiType(Komsai::HEALER) {}
+Komsai::Komsai() : komsaiY(0.0f), komsaiX(0.0f), komsaiSpeed(5.0f), komsaiType(Komsai::HEALER) {}
 
 //Getter Implementations
 int Komsai::get_KomsaiY() const{
-    return komsaiY;
+    return static_cast<int>(komsaiY);
 }
 int Komsai::get_KomsaiX() const{
-    return komsaiX;
+    return static_cast<int>(komsaiX);
 }
-int Komsai::get_KomsaiSpeed() const{
+float Komsai::get_KomsaiSpeed() const{
     return komsaiSpeed;
 }
 
@@ -18,13 +18,13 @@ int Komsai::get_KomsaiType() const{
 }
 
 //Setter Implementations
-void Komsai::set_KomsaiY(int y) {
+void Komsai::set_KomsaiY(float y) {
     this->komsaiY = y;
 }
-void Komsai::set_KomsaiX(int x) {
+void Komsai::set_KomsaiX(float x) {
     this->komsaiX = x;
 }
-void Komsai::set_KomsaiSpeed(int speed) {
+void Komsai::set_KomsaiSpeed(float speed) {
     this->komsaiSpeed = speed;
 }
 void Komsai::set_KomsaiType(KomsaiType type) {
